lambda/requirements.txt
python-dotenv==1.0.0
supabase==2.25.0
openai==2.9.0
elevenlabs==2.26.1
